requirements.txt
pdfminer.six
scikit-learn
joblib
networkx
matplotlib
tqdm
langdetect
numpy 
